% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathteststan-package.R
\docType{package}
\name{breathteststan-package}
\alias{breathteststan-package}
\alias{breathteststan}
\title{breathteststan}
\description{
Stan fits to 13C breath test curves. This
  is an add-on package to package breathtestcore (https://github.com/dmenne/breathtestcore)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dmenne/breathteststan}
  \item \url{https://dmenne.github.io/breathteststan/}
  \item Report bugs at \url{https://github.com/dmenne/breathteststan/issues}
}

}
\author{
\strong{Maintainer}: Dieter Menne \email{dieter.menne@menne-biomed.de}

Authors:
\itemize{
  \item Dieter Menne \email{dieter.menne@menne-biomed.de}
}

Other contributors:
\itemize{
  \item Menne Biomed Consulting Tuebingen [copyright holder]
  \item Benjamin Misselwitz [funder]
  \item Mark Fox [funder]
  \item University Hospital of Zurich, Dep. Gastroenterology [funder, data contributor]
}

}
