% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_diff_by_group.R
\name{coef_diff_by_group.breathteststangroupfit}
\alias{coef_diff_by_group.breathteststangroupfit}
\title{Tabulates breath test parameter differences of groups from Stan group fit}
\usage{
\method{coef_diff_by_group}{breathteststangroupfit}(fit, mcp_group = NULL,
  reference_group = NULL, ...)
}
\arguments{
\item{fit}{Object of class \code{breathtestfit}, for example from
\code{\link{nlme_fit}}, \code{\link{nls_fit}} or \code{\link[breathteststan]{stan_fit}}}

\item{mcp_group}{Not used, always all pairs are compared}

\item{reference_group}{Not used}

\item{...}{Not used}
}
\description{
Given a Stan fit with grouping to 13C breath test curves,
computes point estimated and Bayesian credible intervals for all group pair
differences, for examples of the half emptying time \code{t50}.
}
\examples{
library(dplyr)
data("usz_13c")
data = usz_13c \%>\%
  dplyr::filter( patient_id \%in\%
    c("norm_001", "norm_002", "norm_003", "norm_004", "pat_001", "pat_002","pat_003")) \%>\%
  cleanup_data()
fit = stan_group_fit(data)
coef_diff_by_group(fit)
\dontrun{
fit = nlme_fit(data)
coef_diff_by_group(fit)
}
}
