% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma.R
\name{sigma.breathteststanfit}
\alias{sigma.breathteststanfit}
\title{S3 method to extract the residual standard deviation}
\usage{
\method{sigma}{breathteststanfit}(object, ...)
}
\arguments{
\item{object}{A Stan-based fit returning an S6 method}

\item{...}{Not used}
}
\value{
A numeric value giving the sigma (= average residual standard deviation) term
from the Stan fit.
}
\description{
Functions for S3 method defined in breathtestcore for
\code{CmdStanFit}.
}
