% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breathteststan-package.R
\name{breathteststan-package}
\alias{breathteststan-package}
\alias{breathteststan}
\title{breathteststan}
\description{
Stan fits to 13C breath test curves. This
  is an add-on package to package breathtestcore (https://github.com/dmenne/breathtestcore)
}
